% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrArgos-package.R
\docType{package}
\name{amlrArgos-package}
\alias{amlrArgos-package}
\alias{amlrArgos}
\title{todo}
\description{
Process Argos data for the SWFSC/AERD
}
\details{
Process Argos data for the SWFSC/AERD
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}

Jefferson Hinke \email{jefferson.hinke@noaa.gov}
}
\keyword{package}
