% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrgos-package.R
\docType{package}
\name{amlrgos-package}
\alias{amlrgos-package}
\alias{amlrgos}
\title{Processing AERD Argos Data}
\description{
Process Argos data for the SWFSC/AERD
}
\details{
Process Argos data for the SWFSC/AERD
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}

Jefferson Hinke \email{jefferson.hinke@noaa.gov}
}
\keyword{package}
