% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_info.R
\name{assign_info}
\alias{assign_info}
\title{Import, process, and export data from Argos downloads}
\usage{
assign_info(tt.dat, tt.deployed, iter)
}
\arguments{
\item{tt.dat}{data frame; the raw(ish) data from the ARGOS download}

\item{tt.deployed}{data frame; the log containing deployment information for each tag}

\item{iter}{integer; the index of the file being processed. Not used?}
}
\value{
A data frame more informative Argos data
}
\description{
Add specific information to the ARGOS data set produced by \code{\link{import_argos}}
}
\details{
This ugly pile of braces, brackets, fors, ifs, and elses will assign
Spp, Site, Study, etc. designations to each record in the file
for which deployment information is available.
It goes row by row through the giant data set, so expect it to take awhile...

Ready. Set. Go.
}
