% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_argos.R
\name{format_argos}
\alias{format_argos}
\title{Import, process, and export data from Argos downloads}
\usage{
format_argos(tt, ptt.table)
}
\arguments{
\item{tt}{data frame; processed tag data from Argos CSV files}

\item{ptt.table}{data frame;
file specified by \code{log.file} in \code{\link{import_argos}}}
}
\value{
A data frame with the formatted Argos data
}
\description{
Add specific information to the ARGOS data set produced by \code{\link{import_argos}}
}
\details{
This will only append information to records in the individual files
that are newly imported by \code{\link{import_argos}}, i.e. data from
\code{argos.csv.path} and not from \code{argos.processed.file}.
This function extracts the deployment data for the tags that are in the tag data,
and passes everything to \code{\link{assign_info}}.
}
