% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_argos.R
\name{import_argos}
\alias{import_argos}
\title{Import, process, and export data from Argos downloads}
\usage{
import_argos(
  UPLOAD = FALSE,
  FORMAT.ARGOS = TRUE,
  SPEED.FILTER = TRUE,
  argos.csv.path,
  argos.processed.file,
  export.file,
  log.file,
  tz = "GMT"
)
}
\arguments{
\item{UPLOAD}{logical; if TRUE, then an existing, formatted file housed at
the path specified by \code{argos.processed.file} is already available.
Only the newest files will be appended to this existing data set.
If FALSE, all downloads from the ARGOS server that occurred after
1 Jan 1980 will be imported and formatted.
Please contact the package maintainer if this date needs to be changed}

\item{FORMAT.ARGOS}{logical, default is \code{TRUE}
see \code{\link{format_argos}} for more details}

\item{SPEED.FILTER}{logical, default is \code{TRUE}.
see \code{\link{speed_filter}} for more details}

\item{argos.csv.path}{character; the path to the Argos data}

\item{argos.processed.file}{character; the file that contains the already-
processed data to be appended to the newly processed data}

\item{export.file}{character; the file to which to export the data}

\item{log.file}{character; the PTT log file}

\item{tz}{character; time zone to pass to datetime functions. Default is 'GMT'}
}
\value{
A data frame with the processed Argos data
}
\description{
Import, process, and export data from Argos downloads
}
\details{
todo
}
\examples{

\dontrun{
# Run for all new data
x <- import_argos(
  UPLOAD = FALSE, FORMAT.ARGOS = TRUE, SPEED.FILTER = TRUE,
  argos.csv.path = "processing/input",
  argos.processed.file = "",
  export.file = "ArgosData_2022_01_20_02_13_00_output.csv",
  log.file = "PTTlog.csv"
)

# Run with already-processed data
x <- import_argos(
  UPLOAD = FALSE, FORMAT.ARGOS = TRUE, SPEED.FILTER = TRUE,
  argos.csv.path = "processing/input",
  argos.processed.file = "ArgosData_##_output.csv",
  export.file = "ArgosData_2022_01_20_02_13_00_output.csv",
  log.file = "PTTlog.csv"
)
}

}
