% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_filter.R
\name{speed_filter}
\alias{speed_filter}
\title{Apply a speed filter to each deployment in the Argos dataset}
\usage{
speed_filter(tag.data, ptt.table, deployments)
}
\arguments{
\item{tag.data}{data frame; the raw(ish) data from the ARGOS download}

\item{ptt.table}{data frame;
file specified by \code{log.file} in \code{\link{import_argos}}}

\item{deployments}{character; the unique deployment codes}
}
\value{
A data frame without the filtered Argos data
}
\description{
Apply a speed filter to each deployment in the Argos dataset
}
\details{
This function uses the R package \code{argosfilter} to filter the speeds for each tag
\code{\link[argosfilter]{sdafilter}} removes Z codes and keeps all other positions that meet a
threshold speed criteria set in the function call,
but note that it requires 5 or more points to run.
Tracks with fewer than five points will be set to NA
}
