% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_argos.R
\name{import_argos}
\alias{import_argos}
\title{Import, process, and export data from Argos downloads}
\usage{
import_argos(
  UPLOAD = FALSE,
  FORMAT.ARGOS = TRUE,
  SPEED.FILTER = TRUE,
  argos.csv.path,
  argos.processed.file,
  export.file,
  log.file,
  tz = "GMT"
)
}
\arguments{
\item{UPLOAD}{logical; if TRUE, then an existing, formatted file
housed at the path specified (line 13) is already available.
Only the newest files will be appended to the existing data set.
If FALSE, the analyst must specify a date (lines 43-46, below) and all
downloads from the ARGOS server that occurred after the specified date
will be imported and formatted}

\item{FORMAT.ARGOS}{logical}

\item{SPEED.FILTER}{logical}

\item{argos.csv.path}{character; the path to the Argos data}

\item{argos.processed.file}{character; the file that contains the data to import
TODO clarify this}

\item{export.file}{character; the file to which to export the data}

\item{log.file}{character; the PTT log file}

\item{tz}{character; time zone to pass to datetime functions. Default is 'GMT'}
}
\value{
A data frame with the processed Argos data
}
\description{
Import, process, and export data from Argos downloads
}
\details{
todo
}
